PdFiles Documentation

PdFiles Documentation

PdFiles

PdLibc

PdFiles

PdFiles.PdFiles — Module

PdFiles

Implementations of a better Filehandles.

Includes a Read-Only Filehandle, a better Write-Only Filehandle as well as Mmap-Filehandles allowing both Reading and writing (but not appending (yet)).

Main Functions: pdopen and @om_str

PdFiles.PD_BUFSIZE — Constant

const PD_BUFSIZE = 16 * (1024) * (1024)

Default size of the read/write-buffer of PdReadFile and PdWriteFile.

PdFiles.PdFile — Type

abstract type PdFile <: IO end

Abstract Type for my Filehandles, to overwrite some of Julias functions from io.jl that cause allocations.

PdFiles.PdOpenMode — Type

abstract type PdOpenMode end

Types Implemented by @om_str, used in pdopen

PdFiles.PdOpenMode — Method

PdOpenMode(pattern::AbstractString)

Convert pattern into its corresponding subtype of PdOpenMode.

Recommended Usage of this is with the @om_str macro at parse-time.

PdFiles.PdRawFile — Type

struct PdRawFile <: PdFile 
    fd::Cint
end

Raw Filehandle with no Buffer, that calls the C-Functions for Reading, Writing etc.

Used in PdReadFile and PdWriteFile.

PdFiles.PdReadFile — Type

PdReadFile

Buffered Filehandle for more performant Reading.

Opened via pdopen(filename, om"r").

PdFiles.PdWriteFile — Type

PdReadFile

Buffered Filehandle for more performant Reading.

Opened via pdopen(filename, om"r").

PdFiles.pdopen — Function

pdopen(file::AbstractString, om"w"[, bufsize::Integer])

Open a File in Write-Only Mode.

Truncates the File when opening it.

Returns a PdWriteFile.

PdFiles.pdopen — Function

pdopen(file::AbstractString, om"r"[, bufsize::Integer])

Open a File in Read-Only Mode.

Returns a PdReadFile.

PdFiles.pdopen — Method

pdopen(file::AbstractString, openmode::AbstractString, args...)

Open a File.

Which type of Filehandle is returned depends on openmode.

PdFiles.pdopen — Method

pdopen(file::AbstractString, om"mr+", fs::Integer)

Open a file via Mmap in Read-Write Mode.

The file will be truncated to fs Byte before opening it via Mmap.

Appending is currently not possible and will throw an EOFError.

PdFiles.pdopen — Method

pdopen(file::AbstractString, om"mr+")

Open a file via Mmap in Read-Write Mode.

Appending is currently not possible and will throw an EOFError.

PdFiles.pdopen — Method

pdopen(file::AbstractString, om"mr")

Open a file via Mmap in Read-Only Mode.

PdFiles.@om_str — Macro

@om_str "pattern"
om"pattern"

Convert pattern into its corresponding subtype of PdOpenMode at parse-time.

PdLibc

PdFiles.PdLibc — Module

PdLibc

Interface for all the C-Calls and Constants used.

If a C-Function returns an error-code, Base.systemerror is used to confirm the return code, so the User does not need to do any error-checking themselfs.